refactor(pikachu-game): split components, clean unused code, add unlockable gravity mode
