+++ src/views/pikachu-game/composables/usePikachuGame.ts
@@ -1,158 +1,307 @@
-import { computed, ref } from 'vue'
-import { STORAGE_KEY, directions, difficultyOptions, iconSet, sizeOptions, storyLevels } from '../constants'
+import { computed, nextTick, onMounted, onUnmounted, ref, watch } from 'vue'
+import bgmTrack from '../res/BGM/BGM.mp3'
+import {
+  AUTO_RECOVER_RESHUFFLE_LIMIT,
+  DEFAULT_RECORDS,
+  GRAVITY_UNLOCKED_STORAGE_KEY,
+  STORY_GRID_SIZE,
+  STORY_TOTAL_LEVELS,
+  STORAGE_KEY,
+  difficultyOptions,
+  iconSet,
+  gridPresetMap,
+  sizeOptions,
+} from '../constants'
 import type {
-  Direction,
   DisplayCell,
   GameMode,
   GridSize,
+  LeaderboardCategory,
   LeaderboardStep,
+  MainMode,
   Point,
-  RankedMode,
   RecordItem,
-  SearchNode,
   Tile,
 } from '../types'
-
-interface HintResult {
-  first: Tile
-  second: Tile
-  path: Point[]
-}
-
-interface PackedTile {
-  type: number
-  icon: string
-}
-
-const DEFAULT_RECORDS: ReadonlyArray<RecordItem> = [
-  { name: 'Starter A', score: 40, difficulty: 0, mode: 'classic', timeSpent: 0, createdAt: '2026-01-01T00:00:00.000Z' },
-  { name: 'Starter B', score: 55, difficulty: 1, mode: 'classic', timeSpent: 0, createdAt: '2026-01-02T00:00:00.000Z' },
-  { name: 'Starter C', score: 65, difficulty: 1, mode: 'timed', timeSpent: 120, createdAt: '2026-01-03T00:00:00.000Z' },
-]
+import {
+  applyGravityToBoard,
+  buildBoard as buildBoardGrid,
+  findConnectionPath as findBoardConnectionPath,
+  findHintPath as findBoardHintPath,
+  hasAnyValidMove as hasAnyBoardMove,
+  reshuffleVisibleBoard as reshuffleBoardTiles,
+} from './modules/board'
+import { createAudioController } from './modules/audio'
+import { loadRecordsFromStorage, persistRecordsToStorage } from './modules/storage'
 
 export function usePikachuGame() {
-  const appliedSize = ref<GridSize>(10)
+  const appliedSize = ref<GridSize>(STORY_GRID_SIZE)
   const appliedDifficulty = ref(0)
-  const appliedMode = ref<GameMode>('classic')
-  const storyLevelIndex = ref(0)
+  const appliedMode = ref<GameMode>('story')
 
-  const pendingSize = ref<GridSize>(10)
+  const pendingSize = ref<GridSize>('10x10')
   const pendingDifficulty = ref(0)
-  const pendingMode = ref<GameMode>('classic')
+  const pendingMainMode = ref<MainMode>('story')
+  const pendingCustomMode = ref<Exclude<GameMode, 'story' | 'gravity'>>('classic')
 
   const grid = ref<Tile[][]>([])
   const firstSelected = ref<Tile | null>(null)
   const secondSelected = ref<Tile | null>(null)
+  const failedPairIds = ref<number[]>([])
   const flashPath = ref<Point[]>([])
+  const isHintVisible = ref(false)
   const score = ref(0)
   const assistsLeft = ref(0)
   const timeLeft = ref(180)
-  const message = ref('S?n sàng b?t d?u ván m?i.')
+  const storyElapsedSeconds = ref(0)
+  const storyLevel = ref(1)
+  const message = ref('Sẵn sàng bắt đầu ván mới.')
 
   const isPaused = ref(false)
   const isResolvingPair = ref(false)
+  const isLightMode = ref(false)
+  const isMobilePortrait = ref(false)
+  const isSoundOn = ref(true)
+  const isBgmOn = ref(true)
+  const sfxVolume = ref(50)
+  const bgmVolume = ref(50)
 
   const showNewGameModal = ref(false)
   const showSettingsModal = ref(false)
   const showRecordPrompt = ref(false)
   const showLeaderboardModal = ref(false)
+  const showCreditsModal = ref(false)
   const leaderboardStep = ref<LeaderboardStep>('pick')
+  const leaderboardCategory = ref<LeaderboardCategory>('story')
   const leaderboardDifficulty = ref(0)
   const playerName = ref('')
+  const gravityUnlocked = ref(false)
+  const easterMessage = ref('')
+  const isCreditsBoosted = ref(false)
+  const creditsContainerRef = ref<HTMLElement | null>(null)
+  const creditsContentRef = ref<HTMLElement | null>(null)
+  const creditsEasterButtonRef = ref<HTMLElement | null>(null)
+  const creditsOffsetY = ref(0)
+  const isCreditsAutoStopped = ref(false)
 
   const records = ref<RecordItem[]>([])
+  const autoRecoverAttempts = ref(0)
+  const audio = createAudioController(bgmTrack)
 
   let timerId: ReturnType<typeof setInterval> | null = null
+  let creditsRafId: number | null = null
+  let creditsLastFrame = 0
 
-  const boardRows = computed(() => appliedSize.value)
-  const boardCols = computed(() => appliedSize.value)
+  const activeGridPreset = computed(() => gridPresetMap[appliedSize.value])
+  const boardRows = computed(() => activeGridPreset.value.rows)
+  const boardCols = computed(() => activeGridPreset.value.cols)
   const extRows = computed(() => boardRows.value + 2)
   const extCols = computed(() => boardCols.value + 2)
+  const largestBoardEdge = computed(() => Math.max(boardRows.value, boardCols.value))
+  const totalBoardCells = computed(() => boardRows.value * boardCols.value)
 
   const visibleIconsLeft = computed(() => {
-    return grid.value.reduce((acc, row) => acc + row.filter((tile) => tile.kind === 'icon' && tile.isVisible).length, 0)
+    return grid.value.reduce((acc, row) => {
+      return acc + row.filter((tile) => tile.kind === 'icon' && tile.isVisible).length
+    }, 0)
   })
 
   const isCleared = computed(() => visibleIconsLeft.value === 0 && grid.value.length > 0)
+  const isStoryMode = computed(() => appliedMode.value === 'story' || appliedMode.value === 'gravity')
   const isTimeUp = computed(() => appliedMode.value === 'timed' && timeLeft.value <= 0)
   const isGameOver = computed(() => isCleared.value || isTimeUp.value)
 
   const startTimeBySetup = computed(() => {
-    const base = Math.floor((appliedSize.value * appliedSize.value) / 2.15)
-    return Math.max(80, base - appliedDifficulty.value * 8)
-  })
-
-  const timerTickMs = computed(() => {
-    if (appliedMode.value === 'story') return 1000
-    return Math.max(250, 1000 - appliedDifficulty.value * 120)
+    const base = Math.floor(totalBoardCells.value / 2.2)
+    if (appliedMode.value === 'timed') {
+      return Math.max(75, base) * 3
+    }
+    return Math.max(75, base)
   })
 
   const totalAssists = computed(() => {
     const divider = Math.max(1, appliedDifficulty.value)
-    return Math.max(1, Math.floor(appliedSize.value / divider))
+    return Math.max(1, Math.floor(largestBoardEdge.value / divider))
   })
 
   const timeLabel = computed(() => {
-    const minute = Math.floor(Math.max(timeLeft.value, 0) / 60)
-    const second = Math.max(timeLeft.value, 0) % 60
+    const total =
+      appliedMode.value === 'story' || appliedMode.value === 'gravity'
+        ? Math.max(storyElapsedSeconds.value, 0)
+        : Math.max(timeLeft.value, 0)
+    const minute = Math.floor(total / 60)
+    const second = total % 60
     return `${String(minute).padStart(2, '0')}:${String(second).padStart(2, '0')}`
   })
 
   const playedSeconds = computed(() => {
-    if (appliedMode.value === 'timed') return Math.max(0, startTimeBySetup.value - timeLeft.value)
-    if (appliedMode.value === 'story') return timeLeft.value
-    return 0
+    if (appliedMode.value === 'timed') {
+      return Math.max(0, startTimeBySetup.value - timeLeft.value)
+    }
+    if (appliedMode.value === 'classic') {
+      return 0
+    }
+    return storyElapsedSeconds.value
   })
 
   const timeProgress = computed(() => {
-    if (appliedMode.value !== 'timed') return 100
+    if (appliedMode.value !== 'timed') {
+      return 100
+    }
     return Math.max(0, Math.min(100, Math.round((timeLeft.value / startTimeBySetup.value) * 100)))
   })
 
   const cellSizeClass = computed(() => {
-    if (appliedSize.value === 10) return 'h-9 w-9 text-lg sm:h-11 sm:w-11 sm:text-2xl'
-    if (appliedSize.value === 20) return 'h-5 w-5 text-sm sm:h-6 sm:w-6 sm:text-base'
-    return 'h-3.5 w-3.5 text-[10px] sm:h-4 sm:w-4 sm:text-xs'
+    if (largestBoardEdge.value <= 10) {
+      return 'h-12 w-12 text-[28px] sm:h-[53px] sm:w-[53px] sm:text-[31px]'
+    }
+    if (largestBoardEdge.value <= 20) {
+      return 'h-[26px] w-[26px] text-[17px] sm:h-[31px] sm:w-[31px] sm:text-[20px]'
+    }
+    return 'h-5 w-5 text-[13px] sm:h-[22px] sm:w-[22px] sm:text-[15px]'
   })
 
-  const pathPolylinePoints = computed(() => flashPath.value.map((p) => `${p.x + 0.5},${p.y + 0.5}`).join(' '))
+  const portraitCellSizeClass = computed(() => {
+    if (!isMobilePortrait.value) {
+      return cellSizeClass.value
+    }
+
+    if (largestBoardEdge.value <= 10) {
+      return 'h-9 w-9 text-xl'
+    }
+    if (largestBoardEdge.value <= 20) {
+      return 'h-5 w-5 text-sm'
+    }
+    return 'h-4 w-4 text-xs'
+  })
+
+  const pathPolylinePoints = computed(() => {
+    return flashPath.value.map((point) => `${point.x + 0.5},${point.y + 0.5}`).join(' ')
+  })
+
+  const wrapperClass = computed(() => {
+    return isLightMode.value ? 'bg-text-primary text-bg-deep' : 'bg-bg-deep text-text-primary'
+  })
+
+  const surfaceClass = computed(() => {
+    return isLightMode.value
+      ? 'bg-text-primary border-border-default'
+      : 'bg-bg-surface border-border-default'
+  })
+
+  const panelInnerClass = computed(() => {
+    return isLightMode.value
+      ? 'bg-text-primary border-border-default'
+      : 'bg-bg-deep border-border-default'
+  })
+
+  const textMutedClass = computed(() => {
+    return isLightMode.value ? 'text-bg-elevated' : 'text-text-secondary'
+  })
 
   const classicBoard = computed(() => leaderboardBy(leaderboardDifficulty.value, 'classic'))
   const timedBoard = computed(() => leaderboardBy(leaderboardDifficulty.value, 'timed'))
+  const storyBoard = computed(() => {
+    return records.value
+      .filter((item) => item.mode === 'story')
+      .sort((a, b) => a.timeSpent - b.timeSpent || b.score - a.score)
+      .slice(0, 10)
+  })
+  const gravityBoard = computed(() => {
+    return records.value
+      .filter((item) => item.mode === 'gravity')
+      .sort((a, b) => a.timeSpent - b.timeSpent || b.score - a.score)
+      .slice(0, 10)
+  })
+  const creditsTransformStyle = computed(() => {
+    return { transform: `translateY(${creditsOffsetY.value}px)` }
+  })
+
+  function storyDifficultyByLevel(level: number): number {
+    const ratio = (level - 1) / Math.max(1, STORY_TOTAL_LEVELS - 1)
+    return Math.max(0, Math.min(8, Math.round(ratio * 8)))
+  }
+
+  function getTile(y: number, x: number): Tile | null {
+    const row = grid.value[y]
+    if (!row) {
+      return null
+    }
+    return row[x] ?? null
+  }
 
   const displayCells = computed<DisplayCell[]>(() => {
     const cells: DisplayCell[] = []
     for (let y = 0; y < extRows.value; y++) {
       for (let x = 0; x < extCols.value; x++) {
-        const isOuter = y === 0 || y === extRows.value - 1 || x === 0 || x === extCols.value - 1
-        cells.push({ key: `${x}-${y}`, x, y, tile: isOuter ? null : getTile(y - 1, x - 1), isOuter })
+        const isOuter =
+          y === 0 || y === extRows.value - 1 || x === 0 || x === extCols.value - 1
+        cells.push({
+          key: `${x}-${y}`,
+          x,
+          y,
+          tile: isOuter ? null : getTile(y - 1, x - 1),
+          isOuter,
+        })
       }
     }
     return cells
   })
 
   function loadRecords(): void {
-    try {
-      const cached = localStorage.getItem(STORAGE_KEY)
-      if (cached) {
-        records.value = JSON.parse(cached) as RecordItem[]
-        return
-      }
-    } catch {
-      // fallback to seed data
-    }
-    records.value = [...DEFAULT_RECORDS]
+    records.value = loadRecordsFromStorage(STORAGE_KEY, DEFAULT_RECORDS)
+    persistRecords()
   }
 
   function persistRecords(): void {
-    localStorage.setItem(STORAGE_KEY, JSON.stringify(records.value))
+    persistRecordsToStorage(STORAGE_KEY, records.value)
   }
 
-  function leaderboardBy(difficulty: number, mode: RankedMode): RecordItem[] {
+  function loadGravityUnlockState(): void {
+    gravityUnlocked.value = localStorage.getItem(GRAVITY_UNLOCKED_STORAGE_KEY) === '1'
+  }
+
+  function persistGravityUnlockState(): void {
+    localStorage.setItem(GRAVITY_UNLOCKED_STORAGE_KEY, gravityUnlocked.value ? '1' : '0')
+  }
+
+  function leaderboardBy(difficulty: number, mode: 'classic' | 'timed'): RecordItem[] {
     return records.value
       .filter((item) => item.difficulty === difficulty && item.mode === mode)
       .sort((a, b) => b.score - a.score || a.timeSpent - b.timeSpent)
       .slice(0, 10)
+  }
+
+  function playSelectSound(): void {
+    audio.playSelectSound(isSoundOn.value, sfxVolume.value)
+  }
+
+  function playMatchSound(): void {
+    audio.playMatchSound(isSoundOn.value, sfxVolume.value)
+  }
+
+  function playFailSound(): void {
+    audio.playFailSound(isSoundOn.value, sfxVolume.value)
+  }
+
+  function playHintSound(): void {
+    audio.playHintSound(isSoundOn.value, sfxVolume.value)
+  }
+
+  function playPauseSound(paused: boolean): void {
+    audio.playPauseSound(isSoundOn.value, sfxVolume.value, paused)
+  }
+
+  function playWinSound(): void {
+    audio.playWinSound(isSoundOn.value, sfxVolume.value)
+  }
+
+  function syncBgmState(): void {
+    audio.syncBgmState(isBgmOn.value, bgmVolume.value)
+  }
+
+  async function ensureBgmPlayback(): Promise<void> {
+    await audio.ensureBgmPlayback(isBgmOn.value, bgmVolume.value)
   }
 
   function stopTimer(): void {
@@ -164,374 +313,130 @@ export function usePikachuGame() {
 
   function startTimerIfNeeded(): void {
     stopTimer()
-    if (appliedMode.value === 'classic' || isPaused.value || isGameOver.value) return
+    if (isPaused.value || isGameOver.value) {
+      return
+    }
 
     timerId = setInterval(() => {
-      if (isPaused.value || isGameOver.value) return
+      if (isPaused.value || isGameOver.value) {
+        return
+      }
 
       if (appliedMode.value === 'timed') {
         timeLeft.value = Math.max(0, timeLeft.value - 1)
         if (timeLeft.value === 0) {
-          message.value = 'H?t gi?!'
+          message.value = 'Hết giờ!'
           stopTimer()
         }
         return
       }
 
-      timeLeft.value += 1
-    }, timerTickMs.value)
-  }
+      if (appliedMode.value === 'classic') {
+        return
+      }
 
-  function getTile(y: number, x: number): Tile | null {
-    const row = grid.value[y]
-    if (!row) return null
-    return row[x] ?? null
-  }
-
-  function inBounds(x: number, y: number): boolean {
-    return x >= 0 && x < boardCols.value && y >= 0 && y < boardRows.value
-  }
-
-  function inExtendedBounds(x: number, y: number): boolean {
-    return x >= -1 && x <= boardCols.value && y >= -1 && y <= boardRows.value
-  }
-
-  function shuffle<T>(array: T[]): T[] {
-    const result = [...array]
-    for (let i = result.length - 1; i > 0; i--) {
-      const j = Math.floor(Math.random() * (i + 1))
-      const temp = result[i] as T
-      result[i] = result[j] as T
-      result[j] = temp
-    }
-    return result
-  }
-
-  function obstacleCountByDifficulty(size: number, level: number): number {
-    if (level <= 0) return 0
-    const density = Math.min(0.025 + level * 0.012, 0.12)
-    return Math.floor(size * size * density)
-  }
-
-  function iconVarietyByDifficulty(level: number, size: number): number {
-    const base = Math.floor(size / 5) + 4
-    const scaled = base + level * 2
-    return Math.min(iconSet.length, Math.max(4, scaled))
+      storyElapsedSeconds.value += 1
+    }, 1000)
   }
 
   function buildBoard(): void {
-    const size = appliedSize.value
-    const totalCells = size * size
-    let obstacles = Math.min(obstacleCountByDifficulty(size, appliedDifficulty.value), Math.max(0, totalCells - 2))
-    if ((totalCells - obstacles) % 2 !== 0) obstacles = Math.max(0, obstacles - 1)
-
-    let playableCells = totalCells - obstacles
-    if (playableCells < 2) {
-      playableCells = 2
-      obstacles = totalCells - playableCells
-    }
-
-    const pairs = playableCells / 2
-    const iconTypes: number[] = []
-    const iconChoices = iconVarietyByDifficulty(appliedDifficulty.value, size)
-
-    for (let i = 0; i < pairs; i++) {
-      const type = Math.floor(Math.random() * iconChoices)
-      iconTypes.push(type, type)
-    }
-
-    const pool: Array<{ kind: 'icon' | 'wall'; type: number; icon: string }> = []
-    for (let i = 0; i < obstacles; i++) pool.push({ kind: 'wall', type: -1, icon: '' })
-    iconTypes.forEach((type) => pool.push({ kind: 'icon', type, icon: iconSet[type] as string }))
-
-    const shuffled = shuffle(pool)
-    const next: Tile[][] = []
-    let index = 0
-
-    for (let y = 0; y < size; y++) {
-      const row: Tile[] = []
-      for (let x = 0; x < size; x++) {
-        const item = shuffled[index]
-        row.push({
-          id: index,
-          kind: item?.kind ?? 'icon',
-          type: item?.type ?? -1,
-          icon: item?.icon ?? '',
-          x,
-          y,
-          isVisible: true,
-        })
-        index++
-      }
-      next.push(row)
-    }
-
-    grid.value = next
-  }
-
-  function isWalkable(x: number, y: number, from: Tile, to: Tile): boolean {
-    if ((x === from.x && y === from.y) || (x === to.x && y === to.y)) return true
-    if (!inBounds(x, y)) return true
-    const tile = getTile(y, x)
-    if (!tile || tile.kind === 'wall') return false
-    return !tile.isVisible
-  }
-
-  function reconstructPath(nodes: SearchNode[], endIndex: number): Point[] {
-    const points: Point[] = []
-    let cursor = endIndex
-    while (cursor >= 0) {
-      const node = nodes[cursor]
-      if (!node) break
-      points.push({ x: node.x + 1, y: node.y + 1 })
-      cursor = node.parent
-    }
-    return points.reverse()
-  }
-
-  function compressPath(points: Point[]): Point[] {
-    if (points.length <= 2) return points
-    const compact: Point[] = [points[0] as Point]
-
-    for (let i = 1; i < points.length - 1; i++) {
-      const prev = points[i - 1] as Point
-      const cur = points[i] as Point
-      const next = points[i + 1] as Point
-      const dx1 = cur.x - prev.x
-      const dy1 = cur.y - prev.y
-      const dx2 = next.x - cur.x
-      const dy2 = next.y - cur.y
-      if (dx1 * dy2 !== dy1 * dx2) compact.push(cur)
-    }
-
-    compact.push(points[points.length - 1] as Point)
-    return compact
+    grid.value = buildBoardGrid(boardRows.value, boardCols.value, appliedDifficulty.value, iconSet)
   }
 
   function findConnectionPath(from: Tile, to: Tile): Point[] | null {
-    if (from.kind !== 'icon' || to.kind !== 'icon' || from.type < 0 || to.type < 0 || from.type !== to.type || from.id === to.id) return null
-
-    const queue: SearchNode[] = [{ x: from.x, y: from.y, dir: -1, turns: 0, parent: -1 }]
-    const visited = new Map<string, number>()
-    let head = 0
-
-    while (head < queue.length) {
-      const current = queue[head]
-      if (!current) break
-      const currentIndex = head
-      head++
-
-      for (let i = 0; i < directions.length; i++) {
-        const dir = i as Direction
-        const vector = directions[dir] as readonly [number, number]
-        const nextTurns = current.dir === -1 || current.dir === dir ? current.turns : current.turns + 1
-        if (nextTurns > 2) continue
-
-        let nx = current.x + vector[0]
-        let ny = current.y + vector[1]
-
-        while (inExtendedBounds(nx, ny) && isWalkable(nx, ny, from, to)) {
-          const key = `${nx},${ny},${dir}`
-          const seenTurns = visited.get(key)
-          if (seenTurns === undefined || seenTurns > nextTurns) {
-            visited.set(key, nextTurns)
-            queue.push({ x: nx, y: ny, dir, turns: nextTurns, parent: currentIndex })
-            const nodeIndex = queue.length - 1
-            if (nx === to.x && ny === to.y) return compressPath(reconstructPath(queue, nodeIndex))
-          }
-          nx += vector[0]
-          ny += vector[1]
-        }
-      }
-    }
-
-    return null
+    return findBoardConnectionPath(from, to, grid.value, boardRows.value, boardCols.value)
   }
 
-  function groupedVisibleIcons(): Map<number, Tile[]> {
-    const grouped = new Map<number, Tile[]>()
-    for (const row of grid.value) {
-      for (const tile of row) {
-        if (tile.kind !== 'icon' || !tile.isVisible || tile.type < 0) continue
-        const list = grouped.get(tile.type)
-        if (list) list.push(tile)
-        else grouped.set(tile.type, [tile])
-      }
-    }
-    return grouped
-  }
-
-  function hasAnyValidMove(): boolean {
-    const grouped = groupedVisibleIcons()
-    for (const [, list] of grouped) {
-      for (let i = 0; i < list.length; i++) {
-        for (let j = i + 1; j < list.length; j++) {
-          const a = list[i] as Tile
-          const b = list[j] as Tile
-          if (findConnectionPath(a, b) !== null) return true
-        }
-      }
-    }
-    return false
-  }
-
-  function findHintPath(): HintResult | null {
-    const grouped = groupedVisibleIcons()
-    for (const [, list] of grouped) {
-      for (let i = 0; i < list.length; i++) {
-        for (let j = i + 1; j < list.length; j++) {
-          const a = list[i] as Tile
-          const b = list[j] as Tile
-          const path = findConnectionPath(a, b)
-          if (path !== null) return { first: a, second: b, path }
-        }
-      }
-    }
-    return null
-  }
-
-  function ensureBoardHasMove(maxRetry = 20): void {
-    let tries = 0
-    while (tries < maxRetry && !hasAnyValidMove()) {
-      buildBoard()
-      tries++
-    }
-  }
-
-  function reshuffleRemainingIcons(maxRetry = 20): boolean {
-    const activeTiles: Tile[] = []
-    const packed: PackedTile[] = []
-
-    for (const row of grid.value) {
-      for (const tile of row) {
-        if (tile.kind === 'icon' && tile.isVisible && tile.type >= 0) {
-          activeTiles.push(tile)
-          packed.push({ type: tile.type, icon: tile.icon })
-        }
-      }
-    }
-
-    if (activeTiles.length < 2) return true
-
-    for (let attempt = 0; attempt < maxRetry; attempt++) {
-      const shuffled = shuffle(packed)
-      for (let i = 0; i < activeTiles.length; i++) {
-        const tile = activeTiles[i] as Tile
-        const item = shuffled[i]
-        tile.type = item?.type ?? tile.type
-        tile.icon = item?.icon ?? tile.icon
-        tile.isVisible = true
-      }
-      if (hasAnyValidMove()) return true
-    }
-
-    return false
-  }
-
-  function recoverBoardIfNoMove(): void {
-    firstSelected.value = null
-    secondSelected.value = null
-    flashPath.value = []
-
-    if (reshuffleRemainingIcons()) {
-      message.value = 'Bàn h?t nu?c di. Ðã t? d?ng xáo l?i.'
+  function applyNewGameSettings(): void {
+    if (pendingMainMode.value === 'gravity' && !gravityUnlocked.value) {
+      message.value = 'Mode Trọng lực đang bị khóa. Hãy tìm Easter Egg để mở khóa.'
       return
     }
 
-    buildBoard()
-    ensureBoardHasMove()
-    message.value = 'Bàn h?t nu?c di. Ðã t? d?ng reset bàn m?i.'
-  }
+    appliedMode.value =
+      pendingMainMode.value === 'story'
+        ? 'story'
+        : pendingMainMode.value === 'gravity'
+          ? 'gravity'
+          : pendingCustomMode.value
 
-  function startStoryNextLevel(): void {
-    const nextIndex = storyLevelIndex.value + 1
-    if (nextIndex >= storyLevels.length) {
-      finishWithWin()
-      return
+    storyLevel.value = 1
+    storyElapsedSeconds.value = 0
+
+    if (pendingMainMode.value === 'story' || pendingMainMode.value === 'gravity') {
+      appliedSize.value = STORY_GRID_SIZE
+      appliedDifficulty.value = storyDifficultyByLevel(1)
+    } else {
+      appliedSize.value = pendingSize.value
+      appliedDifficulty.value = pendingDifficulty.value
     }
 
-    storyLevelIndex.value = nextIndex
-    appliedDifficulty.value = storyLevels[nextIndex] as number
-    assistsLeft.value = Math.max(1, totalAssists.value)
-    firstSelected.value = null
-    secondSelected.value = null
-    flashPath.value = []
-    buildBoard()
-    ensureBoardHasMove()
-    message.value = `Story Level ${nextIndex + 1}/${storyLevels.length}`
-  }
-
-  function startFreshBoard(): void {
     score.value = 0
+    timeLeft.value = startTimeBySetup.value
     assistsLeft.value = Math.max(1, totalAssists.value)
+    message.value =
+      appliedMode.value === 'gravity'
+        ? 'Gravity mode: quân sẽ rơi xuống sau mỗi cặp hợp lệ!'
+        : 'Chọn 2 hình giống nhau!'
     isPaused.value = false
     isResolvingPair.value = false
     firstSelected.value = null
     secondSelected.value = null
     flashPath.value = []
+    isHintVisible.value = false
     showRecordPrompt.value = false
     playerName.value = ''
 
     buildBoard()
-    ensureBoardHasMove()
+    autoRecoverAttempts.value = 0
+    autoRecoverNoMoveIfNeeded()
     startTimerIfNeeded()
-  }
-
-  function applyNewGameSettings(): void {
-    if (pendingMode.value === 'story') {
-      appliedMode.value = 'story'
-      appliedSize.value = 20
-      storyLevelIndex.value = 0
-      appliedDifficulty.value = storyLevels[0] as number
-      timeLeft.value = 0
-      message.value = 'Story Level 1 b?t d?u!'
-    } else {
-      appliedMode.value = pendingMode.value
-      appliedSize.value = pendingSize.value
-      appliedDifficulty.value = pendingDifficulty.value
-      storyLevelIndex.value = 0
-      timeLeft.value = startTimeBySetup.value
-      message.value = 'Ch?n 2 hình gi?ng nhau!'
-    }
-
-    startFreshBoard()
+    void ensureBgmPlayback()
     showNewGameModal.value = false
   }
 
-  function applyCurrentSettings(): void {
-    if (pendingMode.value === 'story') {
-      pendingSize.value = 20
-    }
-    appliedMode.value = pendingMode.value
-    appliedSize.value = pendingMode.value === 'story' ? 20 : pendingSize.value
-    appliedDifficulty.value = pendingDifficulty.value
-    if (appliedMode.value === 'story') {
-      storyLevelIndex.value = 0
-      timeLeft.value = 0
-    } else {
-      timeLeft.value = startTimeBySetup.value
-    }
-    message.value = 'Ðã áp d?ng setting. B?t d?u ván m?i.'
-    startFreshBoard()
-    showSettingsModal.value = false
-  }
-
   function togglePause(): void {
-    if (isGameOver.value) return
+    if (isGameOver.value) {
+      return
+    }
+
     isPaused.value = !isPaused.value
     if (isPaused.value) {
       stopTimer()
-      message.value = 'Ðã t?m d?ng.'
-      return
+      message.value = 'Đã tạm dừng.'
+    } else {
+      message.value = 'Tiếp tục nào!'
+      startTimerIfNeeded()
     }
-    message.value = 'Ti?p t?c nào!'
-    startTimerIfNeeded()
+    void ensureBgmPlayback()
+    playPauseSound(isPaused.value)
   }
 
   function finishWithWin(): void {
     stopTimer()
-    message.value = appliedMode.value === 'story' ? 'Hoàn thành toàn b? Story!' : 'B?n chi?n th?ng!'
+    playWinSound()
+    if (appliedMode.value === 'story' || appliedMode.value === 'gravity') {
+      const modeLabel = appliedMode.value === 'gravity' ? 'Gravity' : 'Story'
+      message.value = `Hoàn thành ${modeLabel} ${STORY_TOTAL_LEVELS} level trong ${timeLabel.value}.`
+    } else {
+      message.value = 'Bạn chiến thắng!'
+    }
     showRecordPrompt.value = true
+  }
+
+  function advanceStoryLevel(): void {
+    const nextLevel = storyLevel.value + 1
+    storyLevel.value = nextLevel
+    appliedDifficulty.value = storyDifficultyByLevel(nextLevel)
+    assistsLeft.value = Math.max(1, totalAssists.value)
+    firstSelected.value = null
+    secondSelected.value = null
+    flashPath.value = []
+    isResolvingPair.value = false
+    buildBoard()
+    autoRecoverAttempts.value = 0
+    autoRecoverNoMoveIfNeeded()
+    message.value = `Qua level ${nextLevel - 1}. Sang level ${nextLevel}/${STORY_TOTAL_LEVELS}.`
   }
 
   function resolveMatchedPair(first: Tile, second: Tile, path: Point[]): void {
@@ -547,30 +452,45 @@ export function usePikachuGame() {
       second.type = -1
       second.icon = ''
 
+      if (appliedMode.value === 'gravity') {
+        applyGravityToBoard(grid.value)
+      }
+
       score.value += 10
       flashPath.value = []
       isResolvingPair.value = false
 
       if (isCleared.value) {
-        if (appliedMode.value === 'story') startStoryNextLevel()
-        else finishWithWin()
-        return
-      }
-
-      if (!hasAnyValidMove()) {
-        recoverBoardIfNoMove()
+        if (
+          (appliedMode.value === 'story' || appliedMode.value === 'gravity') &&
+          storyLevel.value < STORY_TOTAL_LEVELS
+        ) {
+          advanceStoryLevel()
+        } else {
+          finishWithWin()
+        }
       } else {
-        message.value = 'N?i thành công!'
+        autoRecoverAttempts.value = 0
+        autoRecoverNoMoveIfNeeded()
+        if (!message.value.includes('Bàn kẹt')) {
+          message.value = 'Nối thành công!'
+        }
       }
-    }, 250)
+    }, 260)
   }
 
   function selectTile(tile: Tile): void {
-    if (isPaused.value || isGameOver.value || isResolvingPair.value) return
-    if (tile.kind !== 'icon' || !tile.isVisible || tile.type < 0) return
+    if (isPaused.value || isGameOver.value || isResolvingPair.value) {
+      return
+    }
+    if (tile.kind !== 'icon' || !tile.isVisible || tile.type < 0) {
+      return
+    }
 
     if (!firstSelected.value) {
       firstSelected.value = tile
+      void ensureBgmPlayback()
+      playSelectSound()
       return
     }
 
@@ -579,73 +499,140 @@ export function usePikachuGame() {
       return
     }
 
-    secondSelected.value = tile
-    const path = findConnectionPath(firstSelected.value, secondSelected.value)
+    const first = firstSelected.value
+    const second = tile
+    if (isHintVisible.value) {
+      flashPath.value = []
+      isHintVisible.value = false
+    }
+    const path = findConnectionPath(first, second)
 
     if (path !== null) {
-      resolveMatchedPair(firstSelected.value, secondSelected.value, path)
+      secondSelected.value = second
+      playMatchSound()
+      resolveMatchedPair(first, second, path)
     } else {
-      message.value = 'Không n?i du?c!'
-      if (!hasAnyValidMove()) recoverBoardIfNoMove()
+      message.value = 'Không nối được!'
+      playFailSound()
+      failedPairIds.value = [first.id, second.id]
+      firstSelected.value = null
+      secondSelected.value = null
+      setTimeout(() => {
+        failedPairIds.value = []
+      }, 280)
+      return
     }
 
     setTimeout(() => {
       firstSelected.value = null
       secondSelected.value = null
-    }, 280)
+    }, 300)
   }
 
   function consumeAssist(): boolean {
     if (assistsLeft.value <= 0) {
-      message.value = 'H?t lu?t tr? giúp.'
+      message.value = 'Hết lượt trợ giúp.'
       return false
     }
     assistsLeft.value -= 1
     return true
   }
 
-  function useHint(): void {
-    if (isPaused.value || isGameOver.value || isResolvingPair.value) return
-    if (!consumeAssist()) return
-
-    const hint = findHintPath()
-    if (!hint) {
-      recoverBoardIfNoMove()
-      return
-    }
-
-    firstSelected.value = hint.first
-    secondSelected.value = hint.second
-    flashPath.value = hint.path
-    message.value = 'G?i ý dã hi?n th?.'
-
-    setTimeout(() => {
-      firstSelected.value = null
-      secondSelected.value = null
-      flashPath.value = []
-    }, 700)
+  function findHintPath(): { first: Tile; second: Tile; path: Point[] } | null {
+    return findBoardHintPath(grid.value)
   }
 
-  function useReload(): void {
-    if (isPaused.value || isGameOver.value || isResolvingPair.value) return
-    if (!consumeAssist()) return
+  function hasAnyValidMove(): boolean {
+    return hasAnyBoardMove(grid.value)
+  }
 
-    if (!reshuffleRemainingIcons()) {
-      buildBoard()
-      ensureBoardHasMove()
-      message.value = 'Ðã reset bàn do không th? xáo an toàn.'
-      return
-    }
-
+  function reshuffleVisibleBoard(): void {
+    reshuffleBoardTiles(grid.value)
     firstSelected.value = null
     secondSelected.value = null
     flashPath.value = []
-    message.value = 'Ðã xáo l?i các icon còn l?i.'
+    isHintVisible.value = false
+  }
+
+  function spawnCurrentLayoutWithFallback(): void {
+    let recovered = false
+    for (let attempt = 0; attempt < 12; attempt++) {
+      buildBoard()
+      if (hasAnyValidMove()) {
+        recovered = true
+        break
+      }
+    }
+
+    if (recovered) {
+      message.value = 'Bàn kẹt quá lâu, đã tạo layout mới cùng màn chơi.'
+    } else {
+      message.value = 'Bàn kẹt, đã tạo layout mới. Hãy dùng tải lại nếu vẫn chưa có nước đi.'
+    }
+  }
+
+  function autoRecoverNoMoveIfNeeded(): void {
+    if (isGameOver.value || isPaused.value || isResolvingPair.value) {
+      return
+    }
+    if (hasAnyValidMove()) {
+      autoRecoverAttempts.value = 0
+      return
+    }
+
+    for (; autoRecoverAttempts.value < AUTO_RECOVER_RESHUFFLE_LIMIT; autoRecoverAttempts.value++) {
+      reshuffleVisibleBoard()
+      if (hasAnyValidMove()) {
+        const attemptCount = autoRecoverAttempts.value + 1
+        autoRecoverAttempts.value = 0
+        message.value = `Bàn kẹt, hệ thống tự tải lại lần ${attemptCount}.`
+        return
+      }
+    }
+
+    autoRecoverAttempts.value = 0
+    spawnCurrentLayoutWithFallback()
+  }
+
+  function useHint(): void {
+    if (isPaused.value || isGameOver.value || isResolvingPair.value) {
+      return
+    }
+    if (!consumeAssist()) {
+      return
+    }
+
+    const hint = findHintPath()
+    if (!hint) {
+      message.value = 'Không có đường đi an toàn.'
+      return
+    }
+
+    flashPath.value = hint.path
+    isHintVisible.value = true
+    message.value = 'Gợi ý đã hiển thị.'
+    void ensureBgmPlayback()
+    playHintSound()
+  }
+
+  function useReload(): void {
+    if (isPaused.value || isGameOver.value || isResolvingPair.value) {
+      return
+    }
+    if (!consumeAssist()) {
+      return
+    }
+
+    autoRecoverAttempts.value = 0
+    reshuffleVisibleBoard()
+    message.value = 'Đã tải lại bàn chơi.'
   }
 
   function saveRecord(): void {
     const name = playerName.value.trim()
-    if (!name) return
+    if (!name) {
+      return
+    }
 
     const record: RecordItem = {
       name,
@@ -656,109 +643,230 @@ export function usePikachuGame() {
       createdAt: new Date().toISOString(),
     }
 
-    records.value = [...records.value, record]
-      .sort((a, b) => b.score - a.score || a.timeSpent - b.timeSpent)
-      .slice(0, 400)
+    records.value = [...records.value, record].slice(-400)
 
     persistRecords()
     showRecordPrompt.value = false
     playerName.value = ''
-    message.value = 'Ðã luu di?m cao.'
+    message.value = 'Đã lưu điểm cao.'
   }
 
   function openLeaderboardFlow(): void {
     leaderboardDifficulty.value = appliedDifficulty.value
+    leaderboardCategory.value = appliedMode.value === 'gravity' ? 'gravity' : 'story'
     leaderboardStep.value = 'pick'
     showLeaderboardModal.value = true
   }
 
-  function prepareNewGameModal(): void {
-    pendingMode.value = appliedMode.value
-    pendingSize.value = appliedSize.value
-    pendingDifficulty.value = appliedDifficulty.value
+  function unlockEasterEgg(): void {
+    if (!gravityUnlocked.value) {
+      gravityUnlocked.value = true
+      persistGravityUnlockState()
+      easterMessage.value = 'Đã mở khóa Gravity mode. Giờ bạn có thể chọn trong Tạo ván mới.'
+      message.value = 'Bạn đã mở khóa Gravity mode!'
+      return
+    }
+
+    easterMessage.value = 'Gravity mode đã được mở khóa từ trước.'
+  }
+
+  function startCreditsBoost(): void {
+    isCreditsBoosted.value = true
+  }
+
+  function stopCreditsBoost(): void {
+    isCreditsBoosted.value = false
+  }
+
+  function resetCreditsPosition(): void {
+    const container = creditsContainerRef.value
+    if (!container) {
+      return
+    }
+    creditsOffsetY.value = container.clientHeight * 0.35
+  }
+
+  function stopCreditsLoop(): void {
+    if (creditsRafId !== null) {
+      cancelAnimationFrame(creditsRafId)
+      creditsRafId = null
+    }
+    creditsLastFrame = 0
+  }
+
+  function tickCredits(timestamp: number): void {
+    const container = creditsContainerRef.value
+    const content = creditsContentRef.value
+    const easterButton = creditsEasterButtonRef.value
+    if (!showCreditsModal.value || !container || !content) {
+      stopCreditsLoop()
+      return
+    }
+
+    if (creditsLastFrame === 0) {
+      creditsLastFrame = timestamp
+    }
+
+    const delta = (timestamp - creditsLastFrame) / 1000
+    creditsLastFrame = timestamp
+    if (!isCreditsAutoStopped.value) {
+      const speed = isCreditsBoosted.value ? 56 : 28
+      creditsOffsetY.value -= speed * delta
+
+      if (creditsOffsetY.value < -content.clientHeight - 24) {
+        creditsOffsetY.value = container.clientHeight * 0.35
+      }
+
+      if (easterButton) {
+        const buttonTop = creditsOffsetY.value + easterButton.offsetTop
+        const buttonCenter = buttonTop + easterButton.clientHeight / 2
+        if (buttonCenter <= container.clientHeight / 2) {
+          isCreditsAutoStopped.value = true
+          stopCreditsBoost()
+          stopCreditsLoop()
+          return
+        }
+      }
+    }
+
+    creditsRafId = requestAnimationFrame(tickCredits)
+  }
+
+  function startCreditsLoop(): void {
+    if (creditsRafId !== null) {
+      return
+    }
+    creditsRafId = requestAnimationFrame(tickCredits)
+  }
+
+  function updateViewportMode(): void {
+    isMobilePortrait.value = window.innerWidth < 768 && window.innerHeight > window.innerWidth
+  }
+
+  onMounted(() => {
+    loadRecords()
+    loadGravityUnlockState()
     showNewGameModal.value = true
-  }
+    updateViewportMode()
+    window.addEventListener('resize', updateViewportMode)
+    void ensureBgmPlayback()
+  })
 
-  function prepareSettingsModal(): void {
-    pendingMode.value = appliedMode.value
-    pendingSize.value = appliedSize.value
-    pendingDifficulty.value = appliedDifficulty.value
-    showSettingsModal.value = true
-  }
+  watch(
+    () => [isBgmOn.value, bgmVolume.value],
+    () => {
+      syncBgmState()
+      if (isBgmOn.value) {
+        void ensureBgmPlayback()
+        return
+      }
+      audio.pauseBgm()
+    },
+  )
 
-  function closeAllModals(): void {
-    showNewGameModal.value = false
-    showSettingsModal.value = false
-    showRecordPrompt.value = false
-    showLeaderboardModal.value = false
-  }
+  watch(
+    () => showCreditsModal.value,
+    async (opened) => {
+      if (!opened) {
+        stopCreditsLoop()
+        stopCreditsBoost()
+        isCreditsAutoStopped.value = false
+        return
+      }
+
+      await nextTick()
+      resetCreditsPosition()
+      isCreditsAutoStopped.value = false
+      startCreditsLoop()
+    },
+  )
+
+  onUnmounted(() => {
+    stopTimer()
+    stopCreditsLoop()
+    window.removeEventListener('resize', updateViewportMode)
+    audio.dispose()
+  })
 
   return {
-    difficultyOptions,
     sizeOptions,
-    storyLevels,
+    difficultyOptions,
+    STORY_TOTAL_LEVELS,
 
     appliedSize,
     appliedDifficulty,
     appliedMode,
-    storyLevelIndex,
-
     pendingSize,
     pendingDifficulty,
-    pendingMode,
+    pendingMainMode,
+    pendingCustomMode,
 
     grid,
     firstSelected,
     secondSelected,
+    failedPairIds,
     flashPath,
     score,
     assistsLeft,
     timeLeft,
+    storyElapsedSeconds,
+    storyLevel,
     message,
 
     isPaused,
     isResolvingPair,
+    isLightMode,
+    isMobilePortrait,
+    isSoundOn,
+    isBgmOn,
+    sfxVolume,
+    bgmVolume,
 
     showNewGameModal,
     showSettingsModal,
     showRecordPrompt,
     showLeaderboardModal,
+    showCreditsModal,
+
     leaderboardStep,
+    leaderboardCategory,
     leaderboardDifficulty,
     playerName,
+    gravityUnlocked,
+    easterMessage,
 
-    records,
+    creditsContainerRef,
+    creditsContentRef,
+    creditsEasterButtonRef,
 
     extRows,
     extCols,
-    isCleared,
-    isTimeUp,
+    displayCells,
+    isStoryMode,
     isGameOver,
-    startTimeBySetup,
-    timerTickMs,
-    totalAssists,
     timeLabel,
-    playedSeconds,
     timeProgress,
-    cellSizeClass,
+    portraitCellSizeClass,
     pathPolylinePoints,
+    wrapperClass,
+    surfaceClass,
+    panelInnerClass,
+    textMutedClass,
     classicBoard,
     timedBoard,
-    displayCells,
+    storyBoard,
+    gravityBoard,
+    creditsTransformStyle,
 
-    loadRecords,
-    stopTimer,
-    startTimerIfNeeded,
     applyNewGameSettings,
-    applyCurrentSettings,
     togglePause,
     selectTile,
     useHint,
     useReload,
-    saveRecord,
     openLeaderboardFlow,
-    prepareNewGameModal,
-    prepareSettingsModal,
-    closeAllModals,
+    saveRecord,
+    unlockEasterEgg,
+    startCreditsBoost,
+    stopCreditsBoost,
   }
 }
