+++ src/views/pikachu-game/meta.ts
@@ -8,3 +8,4 @@ const meta: PageMeta = {
 }
 
 export default meta
+
